requirements.txt
beautifulsoup4==4.11.1
feedparser==6.0.10
requests==2.28.1
pyyaml==6.0
m3u8==3.3.0
openai==0.27.0
whisper==1.0.0
torch==1.13.1
opencv-python==4.7.0.68
pytest==7.2.1
black==23.1.0
ruff==0.0.41 
